hw2/task1/requirements.txt
torch
torchvision
numpy
scipy
Pillow
PyYAML
matplotlib
